Other changes in the move
